Fix: faster points choice
