+++ cmaze/src/gameboard/algorithms/mod.rs
@@ -8,13 +8,18 @@ use smallvec::SmallVec;
 
 use std::{
     fmt, ops,
-    sync::{Arc, Mutex, MutexGuard, RwLock},
+    sync::{Arc, Mutex},
     thread,
 };
 
 use super::{Cell, CellWall, Maze};
 
-use crate::{array::Array3D, dims::*, game::ProgressComm};
+use crate::{
+    array::Array3D,
+    dims::*,
+    game::ProgressComm,
+    progress::{Flag, Progress, ProgressHandle},
+};
 pub use depth_first_search::DepthFirstSearch;
 pub use rnd_kruskals::RndKruskals;
 
@@ -34,121 +39,6 @@ pub enum GenErrorThreaded {
 
 #[derive(Debug)]
 pub struct StopGenerationError;
-
-#[derive(Clone, Debug)]
-pub struct Flag(Arc<RwLock<bool>>);
-
-impl Flag {
-    pub fn new() -> Self {
-        Flag(Arc::new(RwLock::new(false)))
-    }
-
-    pub fn stop(&self) {
-        *self.0.write().unwrap() = true;
-    }
-
-    pub fn is_stopped(&self) -> bool {
-        *self.0.read().unwrap()
-    }
-}
-
-impl Default for Flag {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Clone)]
-pub struct ProgressHandler {
-    jobs: Arc<Mutex<Vec<ProgressHandle>>>,
-}
-
-impl ProgressHandler {
-    #[allow(clippy::new_without_default)]
-    pub fn new() -> Self {
-        Self {
-            jobs: Arc::new(Mutex::new(Vec::new())),
-        }
-    }
-
-    pub fn add(&self) -> ProgressHandle {
-        let progress = ProgressHandle::new(self.clone());
-        self.jobs.lock().unwrap().push(progress.clone());
-        progress
-    }
-
-    pub fn progress(&self) -> Progress {
-        self.jobs.lock().unwrap().iter().fold(
-            Progress {
-                done: 0,
-                from: 0,
-                is_done: true,
-            },
-            |prog, job| prog.combine(&job.progress.lock().unwrap()),
-        )
-    }
-}
-
-#[derive(Clone)]
-pub struct ProgressHandle {
-    progress: Arc<Mutex<Progress>>,
-    handler: ProgressHandler,
-}
-
-impl ProgressHandle {
-    pub fn new(handler: ProgressHandler) -> Self {
-        Self {
-            progress: Arc::new(Mutex::new(Progress::new_empty())),
-            handler,
-        }
-    }
-
-    pub fn split(&self) -> Self {
-        self.handler.add()
-    }
-
-    pub fn lock(&self) -> MutexGuard<Progress> {
-        self.progress.lock().unwrap()
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct Progress {
-    pub done: usize,
-    pub from: usize,
-    pub is_done: bool,
-}
-
-impl Progress {
-    pub fn new(done: usize, from: usize) -> Self {
-        Self {
-            done,
-            from,
-            is_done: false,
-        }
-    }
-
-    pub fn new_empty() -> Self {
-        Self::new(0, 1)
-    }
-
-    pub fn percent(&self) -> f32 {
-        self.done as f32 / self.from as f32
-    }
-
-    pub fn finish(&mut self) {
-        self.done = self.from;
-        self.is_done = true;
-    }
-
-    pub fn combine(&self, other: &Self) -> Self {
-        Self {
-            done: self.done + other.done,
-            from: self.from + other.from,
-            is_done: self.is_done && other.is_done,
-        }
-    }
-}
 
 #[derive(Debug, Clone)]
 pub struct CellMask(Array3D<bool>);
@@ -187,12 +77,33 @@ impl CellMask {
     }
 
     pub fn random_cell(&self, rng: &mut Random) -> Option<Dims3D> {
-        let enabled = self
-            .0
-            .iter_pos()
-            .filter(|&pos| self[pos])
-            .collect::<Vec<_>>();
-        enabled.choose(rng).copied()
+        // If less then 10% of the cells are enabled, we can collect all of them and choose one,
+        // otherwise we can just choose random cell and check that it's enabled.
+
+        let enabled = self.enabled_count();
+        if enabled < self.0.len() / 10 {
+            let enabled = self
+                .0
+                .iter_pos()
+                .filter(|&pos| self[pos])
+                .collect::<Vec<_>>();
+            enabled.choose(rng).copied()
+        } else if enabled > 0 {
+            loop {
+                let size = self.size();
+                let pos = Dims3D(
+                    rng.gen_range(0..size.0),
+                    rng.gen_range(0..size.1),
+                    rng.gen_range(0..size.2),
+                );
+
+                if self[pos] {
+                    return Some(pos);
+                }
+            }
+        } else {
+            None
+        }
     }
 
     pub fn is_connected(&self) -> bool {
@@ -341,6 +252,8 @@ impl Generator {
         assert!(!points.is_empty());
         assert!(points.clone().into_iter().collect::<HashSet<_>>().len() == points.len());
 
+        progress.lock().from = mask.enabled_count();
+
         let size = mask.size();
         let mut groups = Array3D::new_dims(None, size).unwrap();
 
@@ -355,7 +268,6 @@ impl Generator {
         // one.
 
         let mut cycle = 0usize;
-        progress.lock().from = mask.enabled_count();
 
         loop {
             if groups.all(|group| group.is_some()) {
