+++ cmaze/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod game;
 pub mod gameboard;
 pub mod dims;
-
+pub mod progress;
 pub mod array;
